boardSize is now private, fixed some warnings

+++ Eternity II - Genetic Algorithm/GeneticAlgorithm.cpp
@@ -154,7 +154,7 @@ void GeneticAlgorithm::InitRandomPopulation()
   { // Create initialise population of boards with randomised boards
     Board newBoard;
     BoardManager::GetInstance()->InitFullBoard(&newBoard, startPiece);
-    newBoard.boardID = newVec.size() + 1;
+    newBoard.boardID = (int)newVec.size() + 1;
     newVec.push_back(newBoard);
   }
